Create Rekey_map
persue better performance by utilizing std::map::extract that enable users to change the key of an item. write the class "rekey_map" that always mark an erased key to "unused", and we can reuse this unused item as long as insert is called by change its key and valud.